Double-precision AABBs

+++ src/XKTModel/validateXKTArrayBuffer.js
@@ -67,7 +67,7 @@ function inflate(deflatedData) {
         eachEntityId: pako.inflate(deflatedData.eachEntityId, {to: 'string'}),
         eachEntityPrimitiveInstancesPortion: new Uint32Array(pako.inflate(deflatedData.eachEntityPrimitiveInstancesPortion).buffer),
         eachEntityMatricesPortion: new Uint32Array(pako.inflate(deflatedData.eachEntityMatricesPortion).buffer),
-        eachTileAABB: new Float32Array(pako.inflate(deflatedData.eachTileAABB).buffer),
+        eachTileAABB: new Float64Array(pako.inflate(deflatedData.eachTileAABB).buffer),
         eachTileDecodeMatrix: new Float32Array(pako.inflate(deflatedData.eachTileDecodeMatrix).buffer),
         eachTileEntitiesPortion: new Uint32Array(pako.inflate(deflatedData.eachTileEntitiesPortion).buffer),
     };
